+++ common/core/src/main/java/module-info.java
@@ -19,6 +19,7 @@
 /**
  * Common classes used throughout, such as:<br>
  * <br>
+ * {@link is.codion.common.observable.Observable}<br>
  * {@link is.codion.common.event.Event}<br>
  * {@link is.codion.common.event.EventObserver}<br>
  * {@link is.codion.common.state.State}<br>
@@ -37,6 +38,7 @@ module is.codion.common.core {
 	exports is.codion.common.format;
 	exports is.codion.common.item;
 	exports is.codion.common.logging;
+	exports is.codion.common.observable;
 	exports is.codion.common.property;
 	exports is.codion.common.proxy;
 	exports is.codion.common.resource;